+++ pentobi/Main.cpp
@@ -180,12 +180,11 @@ int mainDesktop()
         if (! args.empty())
             initialFile = args.at(0);
         auto style = QQuickStyle::name();
-        // Fusion is broken on Fedora 31 (QTBUG-77107)
-        //if (style.isEmpty() && isDesktop)
-        //{
-        //    style = QStringLiteral("Fusion");
-        //    QQuickStyle::setStyle(style);
-        //}
+        if (style.isEmpty() && isDesktop)
+        {
+            style = QStringLiteral("Fusion");
+            QQuickStyle::setStyle(style);
+        }
         QQmlApplicationEngine engine;
         engine.addImageProvider(QStringLiteral("pentobi"), new ImageProvider);
         auto ctx = engine.rootContext();