+++ pentobi/qml/Main.qml
@@ -31,10 +31,7 @@ ApplicationWindow {
 
     property bool isAndroid: Qt.platform.os === "android"
 
-    // "system" theme would be a better default on desktop with Fusion style
-    // but we currently don't use Fusion (see comment in Main.cpp about
-    // QTBUG-77107)
-    property string themeName: isAndroid ? "dark" : "light"
+    property string themeName: isAndroid ? "dark" : "system"
 
     property QtObject theme: Logic.createTheme(themeName)
     property url folder: androidUtils.getDefaultFolder()
@@ -115,10 +112,7 @@ ApplicationWindow {
         property int visibility
         property alias folder: rootWindow.folder
         property alias displayName: rootWindow.displayName
-        // Changed alias name from themeName to theme to ignore old saved
-        // settings from Pentobi <=17.1 after changing default theme on desktop
-        // (see comment in Main.cpp about QTBUG-77107)
-        property alias theme: rootWindow.themeName
+        property alias themeName: rootWindow.themeName
         property alias exportImageWidth: rootWindow.exportImageWidth
         property alias showToolBar: rootWindow.showToolBar
         property alias showVariations: gameModel.showVariations